Refatoração da contas com menores saldos  para ServiceLayer
